Function mlir_sys::mlirAffineMapPermutationGet

pub unsafe extern "C" fn mlirAffineMapPermutationGet(
    ctx: MlirContext,
    size: isize,
    permutation: *mut c_uint
) -> MlirAffineMap

Creates an affine map with a permutation expression and its size in the context. The permutation expression is a non-empty vector of integers. The elements of the permutation vector must be continuous from 0 and cannot be repeated (i.e. [1,2,0] is a valid permutation. [2,0] or [1,1,2] is an invalid invalid permutation.) The affine map is owned by the context.