Function mlir_sys::mlirBlockArgumentGetOwner

pub unsafe extern "C" fn mlirBlockArgumentGetOwner(
    value: MlirValue
) -> MlirBlock

Returns the block in which this value is defined as an argument. Asserts if the value is not a block argument.