Function mlir_sys::mlirIntegerSetReplaceGet

pub unsafe extern "C" fn mlirIntegerSetReplaceGet(
    set: MlirIntegerSet,
    dimReplacements: *const MlirAffineExpr,
    symbolReplacements: *const MlirAffineExpr,
    numResultDims: isize,
    numResultSymbols: isize
) -> MlirIntegerSet

Gets or creates a new integer set in which the values and dimensions of the given set are replaced with the given affine expressions. dimReplacements and symbolReplacements are expected to point to at least as many consecutive expressions as the given set has dimensions and symbols, respectively. The new set will have numResultDims and numResultSymbols dimensions and symbols, respectively.