Function mlir_sys::mlirSymbolTableReplaceAllSymbolUses

pub unsafe extern "C" fn mlirSymbolTableReplaceAllSymbolUses(
    oldSymbol: MlirStringRef,
    newSymbol: MlirStringRef,
    from: MlirOperation
) -> MlirLogicalResult

Attempt to replace all uses that are nested within the given operation of the given symbol ‘oldSymbol’ with the provided ‘newSymbol’. This does not traverse into nested symbol tables. Will fail atomically if there are any unknown operations that may be potential symbol tables.