Function mlir_sys::mlirSymbolTableWalkSymbolTables

pub unsafe extern "C" fn mlirSymbolTableWalkSymbolTables(
    from: MlirOperation,
    allSymUsesVisible: bool,
    callback: Option<unsafe extern "C" fn(arg1: MlirOperation, arg2: bool, userData: *mut c_void)>,
    userData: *mut c_void
)

Walks all symbol table operations nested within, and including, op. For each symbol table operation, the provided callback is invoked with the op and a boolean signifying if the symbols within that symbol table can be treated as if all uses within the IR are visible to the caller. allSymUsesVisible identifies whether all of the symbol uses of symbols within op are visible.