Type Alias mlir_sys::MlirDiagnosticHandler

pub type MlirDiagnosticHandler = Option<unsafe extern "C" fn(arg1: MlirDiagnostic, userData: *mut c_void) -> MlirLogicalResult>;

Diagnostic handler type. Accepts a reference to a diagnostic, which is only guaranteed to be live during the call. The handler is passed the userData that was provided when the handler was attached to a context. If the handler processed the diagnostic completely, it is expected to return success. Otherwise, it is expected to return failure to indicate that other handlers should attempt to process the diagnostic.

Aliased Type

enum MlirDiagnosticHandler {
    None,
    Some(unsafe extern "C" fn(_: MlirDiagnostic, _: *mut c_void) -> MlirLogicalResult),
}

Variants

None

No value.

Some(unsafe extern "C" fn(_: MlirDiagnostic, _: *mut c_void) -> MlirLogicalResult)

Some value of type T.